updated the types and exports for dialob dashboard
