+++ frontend/dialob-dashboard-material/src/index.tsx
@@ -1,17 +1,22 @@
-import { DialobAdmin } from './DialobAdmin'
+/* eslint-disable react-refresh/only-export-components */
+export { LabelChips, CreateDialog, DeleteDialog, TagTableRow, SortField, Spinner, CustomDatePicker } from './components'
 
-interface CsrfShape {
-  'key': string,
-  'value': string
-}
+export { DialobAdmin } from './DialobAdmin';
+export { DialobAdminView } from './DialobAdminView';
 
-interface DialobAdminConfig {
-  dialobApiUrl: string;
-  setLoginRequired: () => void;
-  setTechnicalError: () => void;
-  language: string;
-  csrf: CsrfShape | undefined;
-  tenantId?: string;
-}
+export { messages } from './intl';
 
-export { DialobAdmin, type DialobAdminConfig }
+export type {
+  DialobAdminViewProps,
+  DialobAdminConfig,
+  CsrfShape,
+  FormConfiguration,
+  FormConfigurationFilters,
+  FormConfigurationTag,
+  FormTag,
+  Metadata,
+  DefaultForm,
+  FormData,
+  FormQuestionnaire,
+  FormMetadata,
+} from './types';
